% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_repeats.R
\name{count_cag_repeats}
\alias{count_cag_repeats}
\title{count repeats}
\usage{
count_cag_repeats(i1, min.repeats = 3)
}
\arguments{
\item{i1}{a dataframe containing a column with the name "insert" which has a character array of DNA}

\item{min.repeats}{is the number of consecutive repeats of a sequence needed to declare it a repeat}
}
\value{
the repeated counts for the given sequence
}
\description{
This function
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
TODO fix so that the pattern is dynamic
}
