% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_samples.R
\name{demultiplex_samples}
\alias{demultiplex_samples}
\title{Parse PacBio Reads}
\usage{
demultiplex_samples(
  input,
  out.folder = "results/demultiplex_samples_testing/",
  run.name = "test",
  Read2N = "CTGTCTCTTATACACATCTCCGAGCCCACGAGAC",
  SampleIndexes = NULL,
  Read2N.max.mm = 1,
  sampleid.length = 8,
  sampleid.max.mm = 1,
  min.reads = 1000,
  file.format = "fastq",
  n.cores = parallel::detectCores() - 1
)
}
\arguments{
\item{input}{Path to the input file}

\item{out.folder}{the location of the output file, where you want your results to be saved #TODO do we want this or should we just return the table}

\item{run.name}{\emph{optional} for naming the output files}

\item{Read2N}{element from 10X}

\item{SampleIndexes}{optional used to subset and filter the reads based on if you know what the sample indexes you are expecting}

\item{Read2N.max.mm}{the max allowed mismatch for the read2n element}

\item{sampleid.length}{how long the 10 sample barcode is expected to be}

\item{sampleid.max.mm}{the max mismatch allowed for barcode from the SampleIndexes to be considered a valid sampleID barcode}

\item{min.reads}{the minimum number of reads needed from a sampleid to be considered non - junk}

\item{file.format}{if the file is in fastq or not, fasta?}
}
\description{
This function loads a fastq and parse the reads into its relevant poart. It assumes that the fastq is alread oriented from 5' prime to 3' prime
it searches for any gene between the exon1p1 and exon1p2 section
it extracts the insert, sample ID and barcode ID
}
