% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reads.R
\name{parse_reads}
\alias{parse_reads}
\title{Parse PacBio Reads}
\usage{
parse_reads(
  input,
  out.file,
  plot = FALSE,
  P5element = "AATGATACGGCGACCACCGAGATCTACAC",
  Spacer = "CGCGTCTG",
  Read1N = "TCGTCGGCAGCGTCAGATGTGTATAAGAGACAG",
  Read2N = "CTGTCTCTTATACACATCTCCGAGCCCACGAGAC",
  P7Element = "ATCTCGTATGCCGTCTTCTGCTTG",
  HTTexon1_p1 = "ATGGCGACCCTGGAAAAGCTGATGAAGGCCTTCGAGTCCCTCAAGTCCTTC",
  HTTexon1_p2 =
    "CAACAGCCGCCACCGCCGCCGCCGCCGCCGCCGCCTCCTCAGCTTCCTCAGCCGCCGCCGCAGGCACAGCCGCTGCTGCCTCAGCCGCAGCCGCCCCCGCCGCCGCCCCCGCCGCCACCCGGCCCGGCTGTGGCTGAGGAGCCGCTGCACCGACC",
  SampleIndexes = c("A1", "A2", "A3", "A4", "A5", "A6", "C1", "C2", "C3", "C4", "C5",
    "C6", "C7", "C8"),
  P5element.max.mm = 1,
  Spacer.max.mm = 0,
  Read1N.max.mm = 1,
  Read2N.max.mm = 1,
  P7Element.max.mm = 1,
  HTTexon1_p1.max.mm = 3,
  HTTexon1_p2.max.mm = 6,
  barcode.length = 16,
  barcode.max.mm = 1,
  sampleid.length = 8,
  sampleid.max.mm = 1,
  file.format = "fastq"
)
}
\arguments{
\item{input}{Path to the input file}

\item{out.file}{the location of the output file, where you want your results to be saved #TODO do we want this or should we just return the table}

\item{plot}{boolean of if you want plots to be created along the way to visualize the process}

\item{P5element}{TODO is this still necessary?}

\item{Spacer}{spacer element from 10X}

\item{Read1N}{element from 10X}

\item{Read2N}{element from 10X}

\item{P7Element}{element from 10X}

\item{HTTexon1_p1}{TODO change this name to something more general; desired sequence match start sequence}

\item{HTTexon1_p2}{TODO change this name to something more general; desired sequence match start sequence}

\item{SampleIndexes}{optional used to subset and filter the reads based on if you know what the sample indexes you are expecting}

\item{P5element.max.mm}{the max allowed mismatch for the P5 element TODO still necessary?}

\item{Spacer.max.mm}{the max allowed mismatch for the P5 element TODO still necessary?}

\item{Read1N.max.mm}{the max allowed mismatch for the read1n element}

\item{Read2N.max.mm}{the max allowed mismatch for the read2n element}

\item{P7Element.max.mm}{the max allowed mismatch for the TODO still necessary?}

\item{HTTexon1_p1.max.mm}{the max allowed mismatch for the first part of the desired sequence matched gene #TODO make this not names HTT}

\item{barcode.length}{how long the 10X barcodes are expected to be}

\item{barcode.max.mm}{the max mismatch allowed for barcode from the whitelist to be considered a valid barcode}

\item{sampleid.length}{how long the 10 sample barcode is expected to be}

\item{sampleid.max.mm}{the max mismatch allowed for barcode from the SampleIndexes to be considered a valid sampleID barcode}

\item{file.format}{if the file is in fastq or not, fasta?}
}
\value{
A dataframe containing the insert, barcode ID, sample ID, and direction of the insert
}
\description{
This function loads a fastq and parse the reads into its relevant poart. It assumes that the fastq is alread oriented from 5' prime to 3' prime
it searches for any gene between the exon1p1 and exon1p2 section
it extracts the insert, sample ID and barcode ID
}
